CPHUN-260828-30: preserve uint16 tracking threshold in parser

+++ lib/Epub/Epub/parsers/ChapterHtmlSlimParser.h
@@ -57,7 +57,7 @@ class ChapterHtmlSlimParser {
   bool focusReadingEnabled;
   uint8_t hangingPunctuationLimitPx;
   bool fixedDialogueSpacing;
-  uint8_t letterSpacingLimitPercent;
+  uint16_t letterSpacingLimitPercent;
   const CssParser* cssParser;
   bool embeddedStyle;
   uint8_t imageRendering;
@@ -159,7 +159,7 @@ class ChapterHtmlSlimParser {
       const float lineCompression, const uint8_t extraParagraphSpacing, const uint8_t paragraphAlignment,
       const uint16_t viewportWidth, const uint16_t viewportHeight, const bool hyphenationEnabled,
       const bool softHyphenEnabled, const bool focusReadingEnabled, const uint8_t hangingPunctuationLimitPx,
-      const bool fixedDialogueSpacing, const uint8_t letterSpacingLimitPercent,
+      const bool fixedDialogueSpacing, const uint16_t letterSpacingLimitPercent,
       const std::function<void(std::unique_ptr<Page>, uint16_t, uint16_t, uint32_t)>& completePageFn,
       const bool embeddedStyle, const std::string& contentBase, const std::string& imageBasePath,
       const uint8_t imageRendering = 0, std::vector<std::string> tocAnchors = {},
